RDK-56050 Merging of OCDM, gst-plugins-rdk into Library - Design Reason for change: comp issues Test Procedure: Refer Ticket
Risks: Low

Priority: P1

+++ middleware/InterfacePlayerRDK.cpp
@@ -712,8 +712,8 @@ void MonitorAV( InterfacePlayerRDK *pInterfacePlayerRDK )
 				{ // avoid logging for initial NULL description
 					MW_LOG_MIL( "MonitorAV_%s: %" G_GINT64_FORMAT ",%" G_GINT64_FORMAT ",%d,%lld",
 							   monitorAVState->description,
-							   monitorAVState->av_position[eGST_MEDIATYPE_VIDEO],
-							   monitorAVState->av_position[eGST_MEDIATYPE_AUDIO],
+							   (gint64)monitorAVState->av_position[eGST_MEDIATYPE_VIDEO],
+							   (gint64)monitorAVState->av_position[eGST_MEDIATYPE_AUDIO],
 							   (int)(monitorAVState->av_position[eGST_MEDIATYPE_VIDEO] - monitorAVState->av_position[eGST_MEDIATYPE_AUDIO]),
 							   monitorAVState->tLastSampled - monitorAVState->tLastReported );
 				}
